reorganize code & update enemy system and add enemy config

+++ Assets/Scripts/Core/GameProperty.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [Serializable]
 public class GameProperty
 {
-    public float attack;
+    public int attack;
     public float damageRate;
     public float criticalHitRate;
     public float fireRate;
